Reworked the "log" module, now called "clogging".
Clogging is now just a little wrapper around the built-in logging module.  It tries to be as non-invasive as possible, and merely passes you a logger from logging.  Also, its usage is nearly identical to logging:

mylogger = clogging.getLogger('myloggersname'[, filename='file.log'])

WARNING: there is still one glitch to work out:  when logging to the terminal, there seems to be DOUBLE outputs.  i think the log messages are getting passed up to the root logger or some ancestor.

The elapsed_time decorator has been moved to the "decorate" module.